Change size, index and offset definitions: use size_t and ptrdiff_t
Instead of unsigned long and long. Adapt surrounding code to use
specific types where possible (i.e., PointIdentifier, CellIdentifier)
but to use size_t and ptrdiff_t where intent is less than crystal
clear.

This commit focuses on changes in the Code/Common directory, and
associated changes in other directories required to build and test
the changes. The remaining directories still need a sweep through
to eliminate inappropriate usage of "long" and "unsigned long".

This commit enables the use of image sizes, offsets and indexes
larger than 2GB on Win64.

For 32-bit Windows, and for 32-bit and 64-bit Linux, this change
should largely be a no-op. The types involved correlate to the
new types being used.

I chose using size_t and ptrdiff_t so that these builds would remain
working as is, including client code, maximizing backwards compatibility.

Code modifications in Testing and Examples are only needed for Win64 builds.
The Testing and Examples would build without code modifications on non-Win64
builds.

+++ Code/Common/itkImageRandomNonRepeatingConstIteratorWithIndex.h
@@ -37,9 +37,9 @@ namespace itk
 class NodeOfPermutation
 {
 public:
-  unsigned long m_Priority;
-  unsigned long m_Index;
-  double        m_Value;
+  size_t m_Priority;
+  size_t m_Index;
+  double m_Value;
 
   NodeOfPermutation ()
   {
@@ -68,9 +68,9 @@ public:
   typedef Statistics::MersenneTwisterRandomVariateGenerator::Pointer GeneratorPointer;
   NodeOfPermutation *m_Permutation;
   GeneratorPointer   m_Generator;
-  unsigned long      m_Size;
+  size_t             m_Size;
 
-  RandomPermutation(unsigned long sz)
+  RandomPermutation(size_t sz)
   {
     m_Size = sz;
     m_Permutation = new NodeOfPermutation[m_Size];
@@ -80,7 +80,7 @@ public:
 
   void Dump()
   {
-    for ( unsigned int i = 0; i < m_Size; i++ )
+    for ( size_t i = 0; i < m_Size; i++ )
       {
       std::cout << m_Permutation[i].m_Value << " " << m_Permutation[i].m_Priority
                 << " " << m_Permutation[i].m_Index << ";";
@@ -88,7 +88,7 @@ public:
       }
   }
 
-  void SetPriority(unsigned long i, unsigned long priority)
+  void SetPriority(size_t i, size_t priority)
   {
     if ( i > m_Size )
       {
@@ -102,7 +102,7 @@ public:
 
   void Shuffle()
   {
-    for ( unsigned int i = 0; i < m_Size; i++ )
+    for ( size_t i = 0; i < m_Size; i++ )
       {
       m_Permutation[i].m_Value = m_Generator->GetVariateWithClosedRange (1.0);
       m_Permutation[i].m_Index = i;
@@ -110,7 +110,7 @@ public:
     std::sort(m_Permutation, m_Permutation + m_Size);
   }
 
-  unsigned long operator[](unsigned long i)
+  size_t operator[](size_t i)
   {
     return m_Permutation[i].m_Index;
   }
@@ -288,7 +288,7 @@ public:
   itkStaticConstMacro(ImageDimension, unsigned int, ::itk::GetImageDimension< TImage >::ImageDimension);
 
   /** Image with priorities */
-  typedef itk::Image< unsigned long, itkGetStaticConstMacro(ImageDimension) > PriorityImageType;
+  typedef itk::Image< size_t, itkGetStaticConstMacro(ImageDimension) > PriorityImageType;
 
   /** Set the priority image.  The priority image controls the order
       of the random selection.  Pixels of the same priority will be
@@ -316,9 +316,9 @@ public:
   }
 
   /** Set/Get number of random samples to get from the image region */
-  void SetNumberOfSamples(unsigned long number);
+  void SetNumberOfSamples(size_t number);
 
-  unsigned long GetNumberOfSamples(void) const;
+  size_t GetNumberOfSamples(void) const;
 
   /** Reinitialize the seed of the random number generator  */
   void ReinitializeSeed();
@@ -330,9 +330,9 @@ public:
 private:
   void UpdatePosition();
 
-  unsigned long      m_NumberOfSamplesRequested;
-  unsigned long      m_NumberOfSamplesDone;
-  unsigned long      m_NumberOfPixelsInRegion;
+  size_t             m_NumberOfSamplesRequested;
+  size_t             m_NumberOfSamplesDone;
+  size_t             m_NumberOfPixelsInRegion;
   RandomPermutation *m_Permutation;
 };
 } // end namespace itk
